+++ CHANGELOG.md
@@ -5,9 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ##[]
+### Added
+- User events timeline
 ### Fixed
 - Some of the DataTables tables got a bit dark in dark mode
--
+
+
 ## [4.56]
 ### Added
 - Test for PanelApp panels loading